core: add /evict rpc endpoint
Add an evict command to corectl to remove a node from a Chain Core
cluster.

I decided to put the evict endpoint in the Core handlers (rather than on
raft.Service under the /raft/ prefix) to parallel the other
cluster-related endpoints called by corectl: /init-cluster and
/join-cluster. Also, it takes the node's address rather than the node ID
to be consistent with the other cluster management commands. Currently,
I don't think we expose the node ID anywhere other than in the logs.

On startup in cmd/cored, I added a call to sinkdb.DB.Ping to perform an
empty write to test the raft configuration. I did this to help handle
the scenario where an evicted node boots up without realizing it's
evicted. The other nodes won't necessarily replicate the eviction entry
to the node if they've already moved on and forgotten the evicted node's
address.